+++ internal/storage/postgres/system.go
@@ -34,7 +34,7 @@ func (p *PgStorage) RunMigrations(ctx context.Context) error {
 					user_id integer not null,
 					number varchar(255) not null ,
 					status varchar(255) not null,
-				    accrual integer,
+				    accrual float,
 				    uploaded_at timestamp with time zone, 
 					foreign key (user_id) references users (id),
 					unique (number, user_id)
